- Add username change condition in profile page - Move username validation function to utils
