+++ src/shared/views/ChangeUserName.js
@@ -1,9 +1,10 @@
 // @flow
-import React from 'react';
+import React, { useState } from 'react';
 import { firebaseConnect } from 'react-redux-firebase';
 import { compose } from 'redux';
 import { withTranslation } from 'react-i18next';
 import { View, StyleSheet, Text, TextInput } from 'react-native';
+import { MessageBarManager } from 'react-native-message-bar';
 import {
     COLOR_WHITE,
     COLOR_DEEP_BLUE,
@@ -18,6 +19,7 @@ import {
 import PageHeader from '../common/PageHeader';
 import Button from '../common/Button';
 import type { NavigationProp, TranslationFunction } from '../flow-types';
+import { checkUserNameExists, validateUserName } from '../utils';
 
 const styles = StyleSheet.create({
     changeUserNameScreen: {
@@ -80,29 +82,50 @@ type InjectedProps = {
 type Props = OwnProps & InjectedProps;
 
 function ChangeUserName(props: Props) {
-    const [newUserName, setNewUserName] = React.useState();
+    const [newUserName, setNewUserName] = useState();
+    const [updatePending, setUpdatePending] = useState(false);
 
     const { t, navigation, firebase } = props;
     const userName = firebase.auth().currentUser?.displayName;
-    const [updatePending, setUpdatePending] = React.useState(false);
 
-    const handleConfirmButtonClick = React.useCallback(() => {
-        if ((newUserName?.length ?? 0) >= MIN_USERNAME_LENGTH) {
-            setUpdatePending(true);
-            firebase
-                .updateAuth({ displayName: newUserName })
-                .then(() => firebase.updateProfile({ username: newUserName }))
-                .then(
-                    () => {
-                        setUpdatePending(false);
-                        navigation.goBack();
-                    },
-                    () => {
-                        setUpdatePending(false);
-                        console.error('failed to change update user profile');
-                    },
-                );
+    const handleConfirmButtonClick = React.useCallback(async () => {
+        const isValid = validateUserName(newUserName);
+
+        if (!isValid) {
+            MessageBarManager.showAlert({
+                title: t('signup:errorOnSignup'),
+                message: t('signup:usernameErrorMessage'),
+                alertType: 'error',
+                shouldHideAfterDelay: false,
+            });
+            return;
         }
+
+        const userNameAlreadyExist = await checkUserNameExists(newUserName);
+
+        if (userNameAlreadyExist) {
+            MessageBarManager.showAlert({
+                title: t('signup:errorOnSignup'),
+                message: t('signup:userNameExistError'),
+                alertType: 'error',
+                shouldHideAfterDelay: false,
+            });
+            return;
+        }
+        setUpdatePending(true);
+        firebase
+            .updateAuth({ displayName: newUserName })
+            .then(() => firebase.updateProfile({ username: newUserName }))
+            .then(
+                () => {
+                    setUpdatePending(false);
+                    navigation.goBack();
+                },
+                () => {
+                    setUpdatePending(false);
+                    console.error('failed to change update user profile');
+                },
+            );
     }, [firebase, newUserName]);
 
     return (
